+++ kroxylicious-integration-tests/src/test/java/io/kroxylicious/proxy/SaslInspectionIT.java
@@ -7,41 +7,46 @@
 package io.kroxylicious.proxy;
 
 import java.time.Duration;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Locale;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
-import java.util.function.BiConsumer;
 
 import org.apache.kafka.clients.CommonClientConfigs;
 import org.apache.kafka.clients.admin.DescribeClusterOptions;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.common.config.SaslConfigs;
 import org.apache.kafka.common.errors.SaslAuthenticationException;
 import org.apache.kafka.common.errors.TimeoutException;
 import org.apache.kafka.common.errors.UnsupportedSaslMechanismException;
+import org.apache.kafka.common.header.Headers;
 import org.apache.kafka.common.protocol.ApiKeys;
-import org.apache.kafka.common.serialization.Serdes;
-import org.assertj.core.api.Assertions;
 import org.assertj.core.api.InstanceOfAssertFactory;
+import org.assertj.core.api.ThrowingConsumer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import io.github.nettyplus.leakdetector.junit.NettyLeakDetectorExtension;
 
 import io.kroxylicious.filters.sasl.inspection.SaslInspection;
+import io.kroxylicious.proxy.authentication.UserFactory;
 import io.kroxylicious.proxy.config.ConfigurationBuilder;
 import io.kroxylicious.proxy.config.NamedFilterDefinition;
 import io.kroxylicious.proxy.config.NamedFilterDefinitionBuilder;
+import io.kroxylicious.proxy.internal.subject.DefaultSaslSubjectBuilderService;
+import io.kroxylicious.proxy.internal.subject.PrincipalAdderConf;
 import io.kroxylicious.proxy.testplugins.ClientAuthAwareLawyer;
 import io.kroxylicious.proxy.testplugins.ClientAuthAwareLawyerFilter;
 import io.kroxylicious.proxy.testplugins.ProtocolCounter;
 import io.kroxylicious.proxy.testplugins.ProtocolCounterFilter;
+import io.kroxylicious.test.assertj.HeadersAssert;
 import io.kroxylicious.test.assertj.KafkaAssertions;
-import io.kroxylicious.test.tester.KroxyliciousTester;
 import io.kroxylicious.testing.kafka.api.KafkaCluster;
 import io.kroxylicious.testing.kafka.common.BrokerConfig;
 import io.kroxylicious.testing.kafka.common.SaslMechanism;
@@ -62,7 +67,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 @ExtendWith(KafkaClusterExtension.class)
 @ExtendWith(NettyLeakDetectorExtension.class)
 @SuppressWarnings("java:S5976") // Ignoring 'replace these n tests with a single parameterized one' - we are using the annotated parameters that a parameterized test wouldn't handle nicely.
-class SaslInspectionIT {
+class SaslInspectionIT extends BaseIT {
 
     /**
      * client handshakes with PLAIN
@@ -121,13 +126,19 @@ class SaslInspectionIT {
         String username = "alice";
         String password = "alice-secret";
 
-        assertClientsCanAccessCluster(cluster, null, topic, mechanism,
+        assertClientsCanAccessCluster(cluster, null, null, topic, mechanism,
                 clientLoginModule, username, password,
-                1, 2, Duration.ofMillis(0));
+                1, 2, Duration.ofMillis(0), headers -> {
+                    HeadersAssert.assertThat(headers)
+                            .singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_AUTHORIZATION_ID)
+                            .hasValueEqualTo("alice");
+                    HeadersAssert.assertThat(headers).singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_MECH_NAME)
+                            .hasValueEqualTo(mechanism);
+                });
     }
 
     /**
-     * Same as {@link #shouldAuthenticateWhenSameMechanism_SCRAM_SHA_256(KafkaCluster, Topic)} but with
+     * Same as {@link #shouldProvideSubject(KafkaCluster, Topic)} but with
      * SCRAM-SHA-512 rather than SCRAM-SHA-256.
      */
     @Test
@@ -178,9 +189,15 @@ class SaslInspectionIT {
         String username = "alice";
         String password = "alice-secret";
 
-        assertClientsCanAccessCluster(cluster, Set.of(mechanism), topic, mechanism, clientLoginModule, username, password,
+        assertClientsCanAccessCluster(cluster, Set.of(mechanism), null, topic, mechanism, clientLoginModule, username, password,
                 2, 1,
-                Duration.ofMillis(10_000));
+                Duration.ofMillis(10_000), headers -> {
+                    HeadersAssert.assertThat(headers)
+                            .singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_AUTHORIZATION_ID)
+                            .hasValueEqualTo("alice");
+                    HeadersAssert.assertThat(headers).singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_MECH_NAME)
+                            .hasValueEqualTo(mechanism);
+                });
     }
 
     /**
@@ -198,9 +215,15 @@ class SaslInspectionIT {
         String username = "alice";
         String password = "alice-secret";
 
-        assertClientsCanAccessCluster(cluster, Set.of(mechanism), topic, mechanism, clientLoginModule, username, password,
+        assertClientsCanAccessCluster(cluster, Set.of(mechanism), null, topic, mechanism, clientLoginModule, username, password,
                 2, 2,
-                Duration.ofMillis(10_000));
+                Duration.ofMillis(10_000), headers -> {
+                    HeadersAssert.assertThat(headers)
+                            .singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_AUTHORIZATION_ID)
+                            .hasValueEqualTo("alice");
+                    HeadersAssert.assertThat(headers).singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_MECH_NAME)
+                            .hasValueEqualTo(mechanism);
+                });
     }
 
     /**
@@ -213,7 +236,7 @@ class SaslInspectionIT {
     void shouldNotAuthenticateWhenNoCommonMechanism(@SaslMechanism(value = "PLAIN", principals = {
             @SaslMechanism.Principal(user = "alice", password = "alice-secret") }) KafkaCluster cluster,
                                                     Topic topic) {
-        var config = buildProxyConfig(cluster, null);
+        var config = buildProxyConfig(cluster, null, null);
 
         try (var tester = kroxyliciousTester(config);
                 var producer = tester.producer(Map.of(
@@ -241,7 +264,7 @@ class SaslInspectionIT {
     @Test
     void shouldNotConnectWhenClientNotConfiguredForSasl(@SaslMechanism(value = "PLAIN", principals = {
             @SaslMechanism.Principal(user = "alice", password = "alice-secret") }) KafkaCluster cluster) {
-        var config = buildProxyConfig(cluster, Set.of("PLAIN"));
+        var config = buildProxyConfig(cluster, Set.of("PLAIN"), null);
 
         try (var tester = kroxyliciousTester(config);
                 var admin = tester.admin(Map.of(
@@ -255,9 +278,60 @@ class SaslInspectionIT {
         }
     }
 
+    /**
+     * Verifies that filters actually receive the default subject created by the SaslInspection default subject builder implementation.
+     *
+     * @param cluster cluster
+     * @param topic topic
+     */
+    @Test
+    void shouldProvideSubject(@SaslMechanism(value = "SCRAM-SHA-256", principals = {
+            @SaslMechanism.Principal(user = "alice", password = "alice-secret") }) KafkaCluster cluster,
+                              Topic topic) {
+
+        String mechanism = "SCRAM-SHA-256";
+        String clientLoginModule = "org.apache.kafka.common.security.scram.ScramLoginModule";
+        String username = "alice";
+        String password = "alice-secret";
+
+        assertClientsCanAccessCluster(cluster, topic, mechanism, null, 2, clientLoginModule, username, password, headers -> {
+            HeadersAssert.assertThat(headers)
+                    .singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_AUTHENTICATED_SUBJECT)
+                    .hasValueEqualTo("Subject[principals=[User[name=%s]]]".formatted(username));
+        });
+    }
+
+    /**
+     * Verifies that filters actually receive a customized subject created by a configured subject builder implementation.
+     *
+     * @param cluster cluster
+     * @param topic topic
+     */
+    @Test
+    void shouldProvideSubjectContainingMappedPrincipal(@SaslMechanism(value = "SCRAM-SHA-256", principals = {
+            @SaslMechanism.Principal(user = "alice", password = "alice-secret") }) KafkaCluster cluster,
+                                                       Topic topic) {
+
+        String mechanism = "SCRAM-SHA-256";
+        String clientLoginModule = "org.apache.kafka.common.security.scram.ScramLoginModule";
+        String username = "alice";
+        String password = "alice-secret";
+
+        var subjectBuilderConfig = new DefaultSaslSubjectBuilderService.Config(List.of(
+                new PrincipalAdderConf(DefaultSaslSubjectBuilderService.SASL_AUTHORIZED_ID,
+                        List.of(new io.kroxylicious.proxy.internal.subject.Map("/(.*)/$1/U", null)), UserFactory.class.getName())));
+
+        var expectedUpperCasedUserName = username.toUpperCase(Locale.ROOT);
+        assertClientsCanAccessCluster(cluster, topic, mechanism, subjectBuilderConfig, 2, clientLoginModule, username, password, headers -> {
+            HeadersAssert.assertThat(headers)
+                    .singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_AUTHENTICATED_SUBJECT)
+                    .hasValueEqualTo("Subject[principals=[User[name=%s]]]".formatted(expectedUpperCasedUserName));
+        });
+    }
+
     private static void assertClientsGetSaslAuthenticationException(KafkaCluster cluster, Topic topic, String mechanism, String clientLoginModule, String username,
                                                                     String password) {
-        var config = buildProxyConfig(cluster, Set.of(mechanism));
+        var config = buildProxyConfig(cluster, Set.of(mechanism), null);
 
         String jaasConfig = "%s required%n  username=\"%s\"%n   password=\"%s\";".formatted(clientLoginModule, username, password);
         try (var tester = kroxyliciousTester(config);
@@ -280,14 +354,33 @@ class SaslInspectionIT {
                                                       String clientLoginModule,
                                                       String username,
                                                       String password) {
-        assertClientsCanAccessCluster(cluster, Set.of(mechanism), topic, mechanism,
+        assertClientsCanAccessCluster(cluster, topic, mechanism, null, numAuthReqPerAuth, clientLoginModule, username, password, headers -> {
+            HeadersAssert.assertThat(headers)
+                    .singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_AUTHORIZATION_ID)
+                    .hasValueEqualTo(username);
+            HeadersAssert.assertThat(headers).singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_MECH_NAME)
+                    .hasValueEqualTo(mechanism);
+        });
+    }
+
+    private static void assertClientsCanAccessCluster(KafkaCluster cluster,
+                                                      Topic topic,
+                                                      String mechanism,
+                                                      @Nullable DefaultSaslSubjectBuilderService.Config subjectBuilderConfig,
+                                                      final int numAuthReqPerAuth,
+                                                      String clientLoginModule,
+                                                      String username,
+                                                      String password,
+                                                      ThrowingConsumer<Headers> headersThrowingConsumer) {
+        assertClientsCanAccessCluster(cluster, Set.of(mechanism), subjectBuilderConfig, topic, mechanism,
                 clientLoginModule, username, password,
-                1, numAuthReqPerAuth, Duration.ofMillis(0));
+                1, numAuthReqPerAuth, Duration.ofMillis(0), headersThrowingConsumer);
     }
 
     @SuppressWarnings("java:S2925") // Impossible to integration test reauth without Thread.sleep
     private static void assertClientsCanAccessCluster(KafkaCluster cluster,
                                                       @Nullable Set<String> proxyEnabledSaslMechanisms,
+                                                      @Nullable DefaultSaslSubjectBuilderService.Config subjectBuilderConfig,
                                                       Topic topic,
                                                       String clientSaslMechanism,
                                                       String clientLoginModule,
@@ -295,8 +388,9 @@ class SaslInspectionIT {
                                                       String password,
                                                       final int numBatches,
                                                       final int numAuthReqPerAuth,
-                                                      Duration pauseTime) {
-        var config = buildProxyConfig(cluster, proxyEnabledSaslMechanisms);
+                                                      Duration pauseTime,
+                                                      ThrowingConsumer<Headers> headersAssertion) {
+        var config = buildProxyConfig(cluster, proxyEnabledSaslMechanisms, subjectBuilderConfig);
 
         String jaasConfig = "%s required%n  username=\"%s\"%n   password=\"%s\";".formatted(clientLoginModule, username, password);
         try (var tester = kroxyliciousTester(config)) {
@@ -316,11 +410,12 @@ class SaslInspectionIT {
                     ConsumerConfig.FETCH_MAX_BYTES_CONFIG, "1");
 
             sendReceiveBatches(tester, topic, producerConfig, consumerConfig, numBatches, (batchNum, records) -> {
-                var headers = Assertions.assertThat(records.records(topic.name()))
+                var headers = assertThat(records.records(topic.name()))
                         .as("topic %s records", topic.name())
                         .singleElement()
                         .asInstanceOf(new InstanceOfAssertFactory<>(ConsumerRecord.class, KafkaAssertions::assertThat))
-                        .headers();
+                        .headers()
+                        .satisfies(headersAssertion);
 
                 int newCount = ProtocolCounterFilter.fromBytes(
                         headers.singleHeaderWithKey(ProtocolCounterFilter.requestCountHeaderKey(ApiKeys.SASL_AUTHENTICATE)).value().actual());
@@ -328,11 +423,6 @@ class SaslInspectionIT {
                 assertThat(newCount)
                         .as("Observed number of %s requests @ batch #%s", ApiKeys.SASL_AUTHENTICATE, batchNum)
                         .isEqualTo(numAuthReqPerAuth * batchNum);
-
-                headers.singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_AUTHORIZATION_ID)
-                        .hasValueEqualTo("alice");
-                headers.singleHeaderWithKey(ClientAuthAwareLawyerFilter.HEADER_KEY_CLIENT_SASL_MECH_NAME)
-                        .hasValueEqualTo(clientSaslMechanism);
 
                 if (batchNum < 2 && pauseTime.isPositive()) {
                     try {
@@ -348,52 +438,41 @@ class SaslInspectionIT {
         }
     }
 
-    public static void sendReceiveBatches(KroxyliciousTester tester,
-                                          Topic topic,
-                                          Map<String, Object> producerConfig,
-                                          Map<String, Object> consumerConfig,
-                                          int numBatches,
-                                          BiConsumer<Integer, ConsumerRecords<String, byte[]>> recordConsumer) {
-        try (var producer = tester.producer(producerConfig);
-                var consumer = tester
-                        .consumer(Serdes.String(), Serdes.ByteArray(), consumerConfig)) {
-            int batchNumOneBased = 1;
-            while (batchNumOneBased <= numBatches) {
-                assertThat(producer.send(new ProducerRecord<>(topic.name(), "my-key", "my-value")))
-                        .succeedsWithin(Duration.ofSeconds(5));
-
-                consumer.subscribe(Set.of(topic.name()));
-                var records = consumer.poll(Duration.ofSeconds(10));
-
-                assertThat(records).hasSize(1);
-                recordConsumer.accept(batchNumOneBased, records);
-                batchNumOneBased += 1;
-            }
-        }
-    }
-
-    private static ConfigurationBuilder buildProxyConfig(KafkaCluster cluster, @Nullable Set<String> enableMechanisms) {
-        NamedFilterDefinitionBuilder namedFilterDefinitionBuilder = new NamedFilterDefinitionBuilder(
-                SaslInspection.class.getName(),
-                SaslInspection.class.getName());
-
-        if (enableMechanisms != null) {
-            namedFilterDefinitionBuilder.withConfig("enabledMechanisms", enableMechanisms);
-        }
-        NamedFilterDefinition saslInspection = namedFilterDefinitionBuilder.build();
-        NamedFilterDefinition counter = new NamedFilterDefinitionBuilder(
+    private static ConfigurationBuilder buildProxyConfig(KafkaCluster cluster, @Nullable Set<String> enableMechanisms,
+                                                         @Nullable DefaultSaslSubjectBuilderService.Config subjectBuilderConfig) {
+        var saslInspection = buildSaslInspector(enableMechanisms, subjectBuilderConfig);
+        var counter = new NamedFilterDefinitionBuilder(
                 ProtocolCounter.class.getName(),
                 ProtocolCounter.class.getName())
                 .withConfig(
                         "countRequests", Set.of(ApiKeys.SASL_AUTHENTICATE),
                         "countResponses", Set.of(ApiKeys.SASL_AUTHENTICATE))
                 .build();
-        NamedFilterDefinition lawyer = new NamedFilterDefinitionBuilder(
+        var lawyer = new NamedFilterDefinitionBuilder(
                 ClientAuthAwareLawyer.class.getName(),
                 ClientAuthAwareLawyer.class.getName())
                 .build();
         return proxy(cluster)
                 .addToFilterDefinitions(saslInspection, counter, lawyer)
                 .addToDefaultFilters(saslInspection.name(), counter.name(), lawyer.name());
+    }
+
+    private static NamedFilterDefinition buildSaslInspector(@Nullable Set<String> enableMechanisms,
+                                                            @Nullable DefaultSaslSubjectBuilderService.Config subjectBuilderConfig) {
+        var saslInspectorBuilder = new NamedFilterDefinitionBuilder(
+                SaslInspection.class.getName(),
+                SaslInspection.class.getName());
+
+        var saslInspectorConfig = new HashMap<String, Object>();
+        Optional.ofNullable(enableMechanisms).ifPresent(value -> {
+            saslInspectorConfig.put("enabledMechanisms", value);
+        });
+        Optional.ofNullable(subjectBuilderConfig).ifPresent(value -> {
+            saslInspectorConfig.put("subjectBuilder", DefaultSaslSubjectBuilderService.class.getName());
+            saslInspectorConfig.put("subjectBuilderConfig", subjectBuilderConfig);
+        });
+        saslInspectorBuilder.withConfig(saslInspectorConfig);
+
+        return saslInspectorBuilder.build();
     }
 }